fix: complex task with children are now editable
