+++ src/main/java/com/lask/controller/LaskController.java
@@ -4,21 +4,17 @@ import com.lask.TaskEditApplication;
 import com.lask.model.AbstractTaskFactory;
 import com.lask.model.StdTaskBuilder;
 import com.lask.model.StdTaskFactory;
-import com.lask.model.TaskBuilder;
 import com.lask.model.task.Task;
 import com.lask.model.task.TaskList;
 import com.lask.model.xml.XMLTaskLoader;
-import com.lask.view.TreeItemTaskVisitor;
+import com.lask.view.task.visitor.TreeItemTaskVisitor;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.control.Label;
 import javafx.scene.control.TreeTableView;
 import javafx.scene.layout.VBox;
 import javafx.stage.FileChooser;
-import javafx.stage.Stage;
 
 import java.io.File;
 import java.io.IOException;
